requirements.txt
web3>=6.0.0,<8.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.26.0,<3.0.0
matplotlib>=3.8.0,<4.0.0
requests>=2.31.0,<3.0.0